requirements.txt
asyncpg==0.30.0
aiogram==3.16.0
loguru==0.7.3
python-dotenv==1.0.0
tqdm==4.67.1
httpx==0.28.1
httpx-socks==0.10.0
telethon==1.38.1
tiktoken==0.8.0
supabase==2.11.0